Move all mocks into a common module for unit tests so unit tests don't have to depend on each other.
