+++ tests/mobly/test_runner_test.py
@@ -24,7 +24,7 @@ from mobly import signals
 from mobly import test_runner
 
 from tests import mock_controller
-from tests.mobly.controllers import android_device_test
+from tests import mocks
 
 
 class TestRunnerTest(unittest.TestCase):
@@ -151,13 +151,13 @@ class TestRunnerTest(unittest.TestCase):
         self.assertEqual(tr.results.controller_info, expected_info)
 
     @mock.patch('mobly.controllers.android_device_lib.adb.AdbProxy',
-                return_value=android_device_test.MockAdbProxy(1))
+                return_value=mocks.MockAdbProxy(1))
     @mock.patch('mobly.controllers.android_device_lib.fastboot.FastbootProxy',
-                return_value=android_device_test.MockFastbootProxy(1))
+                return_value=mocks.MockFastbootProxy(1))
     @mock.patch('mobly.controllers.android_device.list_adb_devices',
                 return_value=["1"])
     @mock.patch('mobly.controllers.android_device.get_all_instances',
-                return_value=android_device_test.get_mock_ads(1))
+                return_value=mocks.get_mock_ads(1))
     def test_run_two_test_classes(self, mock_get_all, mock_list_adb,
                                   mock_fastboot, mock_adb,):
         """Verifies that runing more than one test class in one test run works
